Created ExceptionHandler OperationInterceptor to handle outputfor all methods in one place

+++ src/RestfulService/Handlers/ArtistHandler.cs
@@ -1,11 +1,6 @@
 ﻿using System;
-using System.Configuration;
 using System.IO;
-using System.Reflection;
-using log4net;
 using OpenRasta.Web;
-using RestfulService.Exceptions;
-using RestfulService.OperationResults;
 using RestfulService.Resources;
 using RestfulService.Utility.IO.Readers;
 using RestfulService.Utility.IO.Writers;
@@ -15,11 +10,7 @@ namespace RestfulService.Handlers
 	public class ArtistHandler {
 		private readonly IWriter<Artist> _writer;
 		private readonly IReader<Artist> _reader;
-		private readonly ILog _log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
-
-		private readonly string _baseUrl = ConfigurationManager.AppSettings["Application.BaseUrl"];
-		private const string URI_STRING_FORMAT = "{0}artist/{1}";
-
+		
 		public ArtistHandler(IWriter<Artist> writer, IReader<Artist> reader) {
 			_writer = writer;
 			_reader = reader;
@@ -27,71 +18,39 @@ namespace RestfulService.Handlers
 
 		[HttpOperation("GET")]
 		public OperationResult Get(Artist artist) {
-
-			try {
-				Artist fromFile = _reader.ReadFromFile(artist.Id);
-				return new OperationResult.OK(fromFile);
-			}
-			catch (FileNotFoundException) {
-				return new OperationResult.NotFound { Description = String.Format("Artist {0} not found", artist.Id) };
-			}
-			catch (Exception ex) {
-				_log.Error(ex);
-				return new OperationResult.InternalServerError();
-			}
+			Artist fromFile = _reader.ReadFromFile(artist.Id);
+			return new OperationResult.OK(fromFile);
 		}
 
 		[HttpOperation("POST")]
 		public OperationResult Post(Artist artist) {
 
-			var uriString = CreateUriString(artist.Id);
-
-			try {
-				_writer.CreateFile(artist);
-				return new OperationResult.Created { RedirectLocation = new Uri(uriString), ResponseResource = artist };
-			} catch (ResourceExistsException) {
-				return new OperationResult.Found { RedirectLocation = new Uri(uriString) };
-			} catch (Exception ex) {
-				_log.Error(ex);
-				return new OperationResult.InternalServerError();
-			}
+			var uriString = ServiceEnvironment.CreateArtistUriString(artist.Id);
+			_writer.CreateFile(artist);
+			return new OperationResult.Created { RedirectLocation = new Uri(uriString), ResponseResource = artist };
 		}
 
 		[HttpOperation("PUT")]
 		public OperationResult Put(Artist artist) {
 
-			try {
-				var artistToUpdate = _reader.ReadFromFile(artist.Id);
-				ReMapArtist(artist, artistToUpdate);
+			var artistToUpdate = _reader.ReadFromFile(artist.Id);
+			ReMapArtist(artist, artistToUpdate);
 
-				_writer.UpdateFile(artistToUpdate);
-				var uriString = CreateUriString(artistToUpdate.Id);
+			_writer.UpdateFile(artistToUpdate);
+			var uriString = ServiceEnvironment.CreateArtistUriString(artistToUpdate.Id);
 
-				return new OperationResult.NoContent { ResponseResource = artistToUpdate, RedirectLocation = new Uri(uriString) };
-
-			} catch (FileNotFoundException) {
-				return new OperationResult.NotFound();
-			} catch (Exception ex) {
-				_log.Error(ex);
-				return new OperationResult.InternalServerError();
-			}
+			return new OperationResult.NoContent { ResponseResource = artistToUpdate, RedirectLocation = new Uri(uriString) };
 		}
 
 		[HttpOperation("DELETE")]
 		public OperationResult Delete(Artist artist) {
+			string filePath = ServiceEnvironment.GetFilePath(artist.Id, "C:/artist");
+
+			if (!File.Exists(filePath))
+				throw new FileNotFoundException(string.Format("Could not find {0}", filePath));
 			
-			try {
-				_writer.DeleteFile(artist.Id);
-				return new OperationResult.NoContent();
-			} catch (FileNotFoundException) {
-				return new OperationResult.NotFound();
-			} catch (IOException) {
-				var uriString = CreateUriString(artist.Id);
-				return new OperationResult.MethodNotAllowed(new Uri(uriString), HttpVerb.DELETE.ToString(), artist.Id);
-			} catch (Exception ex) {
-				_log.Error(ex);
-				return new ServiceUnavailable();
-			}
+			_writer.DeleteFile(artist.Id);
+			return new OperationResult.NoContent();
 		}
 
 		private static void ReMapArtist(Artist fromArtist, Artist toArtist) {
@@ -99,10 +58,6 @@ namespace RestfulService.Handlers
 				toArtist.Name = fromArtist.Name;
 			if (!string.IsNullOrEmpty(fromArtist.Genre))
 				toArtist.Genre = fromArtist.Genre;
-		}
-
-		private string CreateUriString(int artistId) {
-			return string.Format(URI_STRING_FORMAT, _baseUrl, artistId);
 		}
 	}
 }
